refactor: use named supabase client
