+++ supabase/functions/store-tables.test.ts
@@ -6,7 +6,7 @@ let settingsBuilder: any
 let integrationsBuilder: any
 
 async function loadClient() {
-  supabase = (await import('../supabaseClient.js')).default
+  supabase = (await import('../../shared/supabase/serverClient.ts')).supabase
 }
 
 describe('store tables queries', () => {
