+++ supabase/functions/supabaseClientHeaders.test.ts
@@ -1,12 +1,13 @@
 import { describe, it, expect } from 'vitest'
-import supabase, { DEFAULT_SUPABASE_KEY } from '../supabaseClient.js'
+import { supabase } from '../../shared/supabase/serverClient'
 
 describe('supabase client headers', () => {
   it('includes default authorization headers', () => {
+    const key = process.env.SUPABASE_SERVICE_ROLE_KEY
     expect((supabase as any).headers).toEqual(
       expect.objectContaining({
-        apikey: DEFAULT_SUPABASE_KEY,
-        Authorization: `Bearer ${DEFAULT_SUPABASE_KEY}`,
+        apikey: key,
+        Authorization: `Bearer ${key}`,
       })
     )
   })
